requirements.txt
asgiref==3.8.1
Django==5.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
